store/requirements.txt
alembic==1.11.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.3
defusedxml==0.7.1
Deprecated==1.2.14
Django==3.2.15
django-allauth==0.55.0
django-debug-toolbar==3.6.0
django-environ==0.9.0
django-extensions==3.2.1
django-filter==23.2
django-redis==5.2.0
djangorestframework==3.14.0
exceptiongroup==1.1.2
flake8==5.0.4
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
import-deps==0.2.0
iniconfig==2.0.0
isort==5.10.1
kombu==5.3.1
loguru==0.7.0
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.1
Pillow==9.2.0
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.8.0
pytest==7.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==2.0.15
sqlparse==0.4.4
stripe==4.1.0
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
win32-setctime==1.1.0
wrapt==1.15.0
